% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_state_clusters.R
\name{calc_state_clusters}
\alias{calc_state_clusters}
\title{Identify clusters of adjacent states to mosaic to national raster}
\usage{
calc_state_clusters(state_list, tier = 1, plot_clusters = F, mult = 0.8)
}
\arguments{
\item{state_list}{list of all state rasters}

\item{tier}{numeric indicating round of hierarchical processing (usually 1-3 rounds)}

\item{plot_clusters}{logical, make map indicating cluster membership?}

\item{mult}{numeric, multiplier indicating increase in search distance compared with the previous tier}
}
\description{
Internal function that identifies groups for 'mosaic_states', not to be called directly.
}
\details{
This function is very similar to calc_tile_clusters, but has some different defaults assuming that input rasters will be whole states, not sub-state tiles.
}
