% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tile_clusters.R
\name{calc_tile_clusters}
\alias{calc_tile_clusters}
\title{Identify clusters of adjacent tiles to mosaic to larger raster}
\usage{
calc_tile_clusters(tile_list, chunksize, plot_clusters = F)
}
\arguments{
\item{tile_list}{list of all raster tiles}

\item{chunksize}{numeric indicating how many tiles should be in each cluster}

\item{plot_clusters}{logical, make map indicating cluster membership?}
}
\description{
Internal function that identifies groups for 'mosaic_tiles', not to be called directly
}
\keyword{Internal}
