% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CapStr.R
\name{CapStr}
\alias{CapStr}
\title{Helper function to capitalize a string}
\usage{
CapStr(y)
}
\arguments{
\item{x}{input string}
}
\value{
capitalized output string
}
\description{
Helper function to capitalize a string
}
\keyword{bees}
\keyword{ecology}
\keyword{landscape}
\keyword{spatial}
