% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDL_reclass.R
\name{CDL_reclass}
\alias{CDL_reclass}
\title{Reclassify land use raster}
\usage{
CDL_reclass(rasterpath, reclasstable, from, to, writerast = F, outpath,
  meanreclass)
}
\arguments{
\item{rasterpath}{file path of land use raster file}

\item{reclasstable}{R data frame of reclass table}

\item{from}{column in reclass table with land use raster values (e.g. CDL values from 1 to 254)}

\item{to}{column in reclass table with new values (e.g. pesticide index of each land use)}

\item{writerast}{logical, should write file of reclassified raster?}

\item{outpath}{if writerast is TRUE, path name where reclass raster should be written}

\item{meanreclass}{logical, return mean value of reclassified raster}
}
\description{
Reclassify land use (e.g. Crop Data Layer) raster by supplied matrix with option to calculate mean of new raster
}
\examples{
'CDL_reclass' example forthcoming.
}
\keyword{bees}
\keyword{ecology}
\keyword{landscape}
\keyword{spatial}
