% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDL_reclass_cmpd.R
\name{CDL_reclass_cmpd}
\alias{CDL_reclass_cmpd}
\title{creates a state map of pesticide use for a particular compound by attributing pesticide use intensity (kg/ha) to land use/land cover in the Cropland Data Layer}
\usage{
CDL_reclass_cmpd(state, year, cmpd, from, to, writerast = F, meanreclass)
}
\arguments{
\item{state}{name of state you want to reclassify (e.g. 'MD')}

\item{year}{year of data you want to reclassify (e.g. 2012)}

\item{cmpd}{compound name from the USGS pesticide dataset (e.g. 'glyphosate')}

\item{from}{column in reclass table with land use raster values (e.g. CDL values from 1 to 254)}

\item{to}{column in reclass table with new values (e.g. main current option is application, 'kg_ha' - kg/hectare)}

\item{writerast}{logical, should write file of reclassified raster?}

\item{meanreclass}{logical, return mean value of reclassified raster}

\item{outpath}{if writerast is TRUE, path name where reclass raster should be written}
}
\description{
creates a state map of pesticide use for a particular compound by attributing pesticide use intensity (kg/ha) to land use/land cover in the Cropland Data Layer
}
\details{
This function is a wrapper for beecoSp's 'CDL_reclass' specific to reclassifying state maps into pesticide use intensity.
}
