% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipland.R
\name{clipland}
\alias{clipland}
\title{Clip landscape raster by polygon or buffered point}
\usage{
clipland(rasterpath, featurepath, points = F, buffer = NA)
}
\arguments{
\item{rasterpath}{Full path name for landscape raster file}

\item{featurepath}{Full path name for polygon or point shapefile}

\item{usepoints}{Start with points shapefile of landscape centroids}

\item{bufferdist}{Radius of desired landscape buffer in km}

\item{outdir}{Directory where .tif landscape clips are to be stored, do not include final backslash}
}
\description{
Clip landscape raster by polygon or buffered point features
}
\examples{
clipland()
}
\keyword{bees}
\keyword{ecology}
\keyword{landscape}
\keyword{spatial}
