% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipland.R
\name{clipmask}
\alias{clipmask}
\title{Clip landscape raster by polygon or buffered point}
\usage{
clipmask(land, polygonID, polygons, outdir, idvar)
}
\arguments{
\item{land}{landscape raster, passed from 'execute_landmask'}

\item{polygonID}{id of specific polygon used for clip and mask}

\item{polygons}{SpatialPolygonDataFrame of polygons to clip, passed from 'execute_landmask'}

\item{outdir}{Directory where .tif landscape clips are to be stored, passed from 'execute_landmask'}

\item{idvar}{Identifying variable name within feature shapefile, passed from 'execute_landmask'}
}
\description{
Internal function that clip landscapes within 'execute_landmask', not to be called directly
}
\examples{
see 'execute_landmask' for usage example.
}
\keyword{bees}
\keyword{ecology}
\keyword{landscape}
\keyword{spatial}
