% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliprasterbuffer.R
\name{cliprasterbuffer}
\alias{cliprasterbuffer}
\title{Clip raster tiles to smaller size}
\usage{
cliprasterbuffer(rasterpath, buffercells)
}
\arguments{
\item{rasterpath}{file path to raster file}

\item{buffercells}{vector indicating the number of buffer in ncells in x and y dimension.}
}
\description{
Internal function that clips raster tiles by a specified buffer. Useful when tiles were created with overlap to handle edge effects.
}
