% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassign_NA.R
\name{custom_modal}
\alias{custom_modal}
\title{Mode function that calculates mode of specific raster classes}
\usage{
custom_modal(x, ...)
}
\arguments{
\item{x}{input raster file}
}
\description{
Function to identify the most common raster class of a set of allowed classes (e.g. most common agricultural class)
}
\keyword{bees}
\keyword{ecology}
\keyword{landscape}
\keyword{spatial}
