% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distweight_lulc.R
\name{distweight_lulc}
\alias{distweight_lulc}
\title{Calculate distance weighted land cover}
\usage{
distweight_lulc(land_raster, forage_range)
}
\arguments{
\item{land_raster}{File path to land cover raster (centered on sampling location)}

\item{forage_range}{Mean foraging range of bee community of interest (determines shape of distance weighting curve)}
}
\description{
Calculate distance weighted sum of raster classes, assuming middle pixel is focal cell.
}
\details{
Distance weighting is an exponential decline function with a maximum distance of 'forage_range' *2.
See Lonsdorf et al (2009) for details.
'land_raster' must have a radius greater than 'forage_range'*2, but can be much larger (whole county). The raster used for distance weighting is cropped to radius of 'forage_range' times 2.
}
\examples{
Usage example coming soon.
}
