% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_one_raster.R
\name{grid_one_raster}
\alias{grid_one_raster}
\title{Split regional or national raster into gridded tiles}
\usage{
grid_one_raster(
  rasterpath,
  regionalextent = NA,
  rasterID,
  tile_dir,
  div,
  buffercells = c(0, 0),
  NAvalue,
  writetiles = T,
  ...
)
}
\arguments{
\item{rasterpath}{file path of input raster file}

\item{regionalextent}{optional, vector of state names to crop national raster to region of interest}

\item{rasterID}{text string to identify output tiles (e.g. CDL for NASS Cropland Data Layer)}

\item{div}{division factor specifying the number of tiles in x and y dimensions}

\item{buffercells}{number of cells overlapping between adjacent tiles}

\item{NAvalue}{No data or background value of input raster}

\item{writetiles}{logical, write tiles as directory of .tif files?}
}
\value{
A list of raster tiles.
}
\description{
Split regional or national raster into gridded tiles
}
\details{
NAvalue parameter is used to identify raster tiles that are all background (e.g. areas of open water). These background tiles are excluded from list of output tiles and tile directory.
}
\examples{
example forthcoming.
}
\keyword{bees}
\keyword{ecology}
\keyword{landscape}
\keyword{spatial}
