% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_rasters.R
\name{grid_rasters}
\alias{grid_rasters}
\title{Split two regional or national rasters into gridded tiles}
\usage{
grid_rasters(
  rasterpath,
  rasterID,
  regionalextent = NA,
  div,
  buffercells = c(0, 0),
  NAvalues,
  writetiles = T,
  tiledir,
  verbosewrite = F
)
}
\arguments{
\item{rasterpath}{file paths of input raster files}

\item{rasterID}{text strings to identify output tiles (e.g. CDL for NASS Cropland Data Layer)}

\item{regionalextent}{optional, vector of state names to crop national raster to region of interest}

\item{div}{division factor specifying the number of tiles in x and y dimensions}

\item{buffercells}{number of cells overlapping between adjacent tiles}

\item{NAvalues}{No data or background values of input rasters}

\item{writetiles}{logical, write tiles as directory of .tif files?}

\item{tiledir}{path to directory where output tiles should be saved}
}
\value{
A list of raster tiles. Each element of the list is a pair of raster tiles (one tile for each of the two input raster).
}
\description{
Split two regional or national rasters into gridded tiles
}
\details{
NAvalue parameter is used to identify raster tiles that are all background (e.g. areas of open water). These background tiles are excluded from list of output tiles and tile directory.
}
\keyword{bees}
\keyword{ecology}
\keyword{landscape}
\keyword{spatial}
