% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landcomp.R
\name{landcomp}
\alias{landcomp}
\title{Calculate landscape percent land use, or any categorical raster}
\usage{
landcomp(landdir = T, landfiles, writeoutput = T, outfile, attr_path,
  attr_value, background = F, bgvalues = -1000)
}
\arguments{
\item{landdir}{Logical specifying if 'landfiles' is a directory}

\item{landfiles}{Landscape raster files to process (path to directory or list of files)}

\item{writeoutput}{Logical specifying if results should be written to .csv file}

\item{outfile}{If writeoutput=T, path to .csv file to store results (including .csv filename)}

\item{attr_path}{Path to .csv file of raster attribute table}

\item{attr_value}{Column name in attribute table that specifies raster values}

\item{background}{Logical, specify background value?}

\item{bgvalues}{or NoData values to be excluded from total landscape area}
}
\description{
Calculate landscape percent land use, or any categorical raster
}
\examples{
Usage example coming soon.
}
\keyword{bees}
\keyword{ecology}
\keyword{landscape}
\keyword{spatial}
