% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_states.R
\name{mosaic_states}
\alias{mosaic_states}
\title{Mosaic together state rasters to form a national raster}
\usage{
mosaic_states(statedir, tier, ID, outdir, season = NA, usepackage = "gdal")
}
\arguments{
\item{statedir}{file path to directory that contains state rasters in .tif format}

\item{tier}{numeric indicating round of hierarchical processing (usually 1-3 rounds)}

\item{ID}{variable to identify groups of rasters to process together (e.g. CDLYear). Will be included in file name for output raster.}

\item{outdir}{file path to directory where natinoal rasters should be saved}

\item{season}{if applicable, vector of seasons to process. E.g. spring, summer, fall}

\item{verbose}{logical, print many status messages}
}
\description{
Mosaic together state rasters to form a national raster
}
\details{
Due to large files sizes, output rasters from this function are always compressed. This is different from 'mosaic_tiles' function.
}
