% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_tiles.R
\name{mosaic_tiles}
\alias{mosaic_tiles}
\title{Mosaic together raster tiles to form state-level rasters}
\usage{
mosaic_tiles(
  tiledir,
  chunksize1,
  ID,
  outdir,
  season = NA,
  compress = T,
  verbose = F
)
}
\arguments{
\item{tiledir}{file path to directory that contains raster tiles in .tif format}

\item{chunksize1}{numeric indicating how many tiles should be in each cluster}

\item{ID}{variable to identify groups of rasters to process together (e.g. CDLYear). Will be included in file name for output raster.}

\item{outdir}{file path to directory where state rasters should be saved}

\item{season}{if applicable, vector of seasons to process. E.g. spring, summer, fall}

\item{compress}{logical, should state rasters be compressed?}

\item{verbose}{logical, print many status messages}
}
\description{
Mosaic together raster tiles to form state-level rasters
}
