% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclasstables.R
\name{reclasstables}
\alias{reclasstables}
\title{Create reclass tables for a particular pesticide, state, and year}
\usage{
reclasstables(filepath, states, years, write_reclass, outfolder)
}
\arguments{
\item{filepath}{file path for master pesticide use reclass table}

\item{states}{vector of states (characters) to include}

\item{years}{vector of years (integers) to include}

\item{write_reclass}{logical, write_reclass = T writes reclass tables to file}

\item{outfolder}{folder path to write reclass keys}
}
\description{
Function returns individual dataframes for each state-year reclass table and includes option to write reclass tables to file in an indicated folder.
}
