% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choosepoints.R
\name{subsetplots}
\alias{subsetplots}
\title{Subset sites based on proximity to ensure independence of bee communities}
\usage{
subsetplots(points, cutoff)
}
\arguments{
\item{points}{Spatial points object of sites (sampling locations)}

\item{cutoff}{Distance cutoff for independent samples (typically average foraging range of bee(s))}
}
\description{
Subset sites based on proximity to ensure independence of bee communities
}
\examples{
Usage example coming soon.
}
