% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipland.R
\name{bufferproject}
\alias{bufferproject}
\title{Buffer and reproject input points or polygons}
\usage{
bufferproject(rasterpath, featurepath, bufferdist = NA)
}
\arguments{
\item{rasterpath}{Full path name for landscape raster file}

\item{featurepath}{Full path name for point shapefile}

\item{bufferdist}{Radius of desired landscape buffer in meters}
}
\description{
Clip landscape raster by polygon or buffered point features
}
\examples{
bufferproject()
}
\keyword{bees}
\keyword{ecology}
\keyword{landscape}
\keyword{spatial}
