% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipland.R
\name{clipmask}
\alias{clipmask}
\title{Clip landscape raster by polygon or buffered point}
\usage{
clipmask(land, polygonID, polygons, outdir, idvar)
}
\arguments{
\item{land}{landscape raster, passed from 'clipland' function}

\item{polygonID}{id of specific polygon used for clip and mask}

\item{polygons}{SpatialPolygonDataFrame of polygons to clip, passed from'clipland' function}

\item{outdir}{Directory where .tif landscape clips are to be stored, passed from 'clipland'}

\item{idvar}{Identifying variable name within feature shapefile, passed from 'clipland' function}
}
\description{
Clip landscape raster by polygon or buffered point features
}
\examples{
clipmask()
}
\keyword{bees}
\keyword{ecology}
\keyword{landscape}
\keyword{spatial}
