% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipmask.R
\name{clipmask}
\alias{clipmask}
\title{Clip landscape raster by polygon or buffered point}
\usage{
clipmask(land, polygonID, polygons, outdir, idvar, overrast, na_value = NA)
}
\arguments{
\item{land}{landscape raster, passed from 'execute_landclip'}

\item{polygonID}{id of specific polygon used for clip and mask}

\item{polygons}{SpatialPolygonDataFrame of polygons to clip, passed from 'execute_landclip'}

\item{outdir}{Directory where .tif landscape clips are to be stored, passed from 'execute_landclip'}

\item{idvar}{Identifying variable name within feature shapefile, passed from 'execute_landclip'}

\item{overrast}{Logical, should existing rasters with same filename be overwritten?}

\item{na_value}{Numeric, to assign NA values in raster}
}
\description{
Internal function that clips landscapes within 'execute_landclip', not to be called directly
}
\examples{
see 'execute_landclip' for usage example.
}
\keyword{bees}
\keyword{ecology}
\keyword{landscape}
\keyword{spatial}
