% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landcomp.R
\name{landcomp}
\alias{landcomp}
\title{Calculate landscape percent land use, or any categorical raster}
\usage{
landcomp(landdir = T, landfiles, outfile, writeoutput = T, attr_path)
}
\arguments{
\item{landdir}{Logical specifying if 'landfiles' path is a directory}

\item{landfiles}{Path to landscape raster file (or files) to process}

\item{outfile}{Filename and path to .csv file to store results}

\item{idvar}{Identifying variable name within feature shapefile, passed from 'execute_landclip'}

\item{outdir}{Directory where .tif landscape clips are to be stored, passed from 'execute_landclip'}

\item{overrast}{Logical, should existing rasters with same filename be overwritten?}
}
\description{
Calculate landscape percent land use, or any categorical raster
}
\examples{
see 'execute_landclip' for usage example.
}
\keyword{bees}
\keyword{ecology}
\keyword{landscape}
\keyword{spatial}
