% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choosepoints.R
\name{subsetplots}
\alias{subsetplots}
\title{Subset sites based on proximity}
\usage{
subsetplots(points, cutoff)
}
\arguments{
\item{points}{Spatial points object of sites (sampling locations)}

\item{cutoff}{Distance cutoff for independent samples (typically average foraging range of bee(s))}
}
\description{
Select sites that are separated by some minimum distance to ensure independence of bee communities
}
\details{
Distance cutoff must be specified in the same units as projection of spatial points layer.
}
\examples{
Usage example coming soon.
}
