% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choosepoints.R
\name{subsetplots}
\alias{subsetplots}
\title{Filter sites to random subset based on proximity}
\usage{
subsetplots(points, cutoff)
}
\arguments{
\item{points}{Spatial points object of sites (sampling locations)}

\item{cutoff}{Distance cutoff (in meters) for independent samples (typically average foraging range of bee(s))}
}
\description{
Select sites that are separated by some minimum distance to ensure independence of bee communities
}
\details{
Spatial points must lat/long (not projected). See 'distGeo' function (geosphere package) for details of distance calculation.
}
\examples{
Usage example coming soon.
}
